+++ tests/cli/test_main.py
@@ -798,7 +798,6 @@ class TestCLICommand:
         # The command is just 'bash'
         assert command.args == "bash"
 
-        assert command.cwd == str(steam_app.install_path)
         assert command.shell is True
 
         # Correct environment vars were set